merge

+++ drivers/cpufreq/cpufreq_interactive.c
@@ -144,7 +144,7 @@ static bool powersave_bias;
  * Stay at max freq for at least max_freq_hysteresis before dropping
  * frequency.
  */
-static unsigned long max_freq_hysteresis = 100000;
+static unsigned int max_freq_hysteresis = 100000;
 
 static bool io_is_busy = true;
